refactor: split the FTS planning module by node type
fts.rs had grown to 1790 lines holding every FTS concern at once. Split it into
a directory: one module per node type (leaf, compound, scorer, match_filter)
plus prefetch, builder, rules, and planner. Largest module is now 397 lines.

The five entry points the rest of the planner uses are unchanged; only where
they live moved.
